+++ src/test/java/org/folio/sidecar/integration/keycloak/filter/KeycloakJwtFilterTest.java
@@ -25,6 +25,7 @@ import static org.mockito.Mockito.when;
 
 import io.quarkus.security.UnauthorizedException;
 import io.smallrye.jwt.auth.principal.ParseException;
+import io.vertx.core.Future;
 import io.vertx.core.MultiMap;
 import io.vertx.core.http.HttpServerRequest;
 import io.vertx.core.http.impl.headers.HeadersMultiMap;
@@ -35,7 +36,7 @@ import java.util.Map;
 import java.util.UUID;
 import java.util.function.Consumer;
 import org.eclipse.microprofile.jwt.JsonWebToken;
-import org.folio.jwt.openid.JsonWebTokenParser;
+import org.folio.sidecar.integration.keycloak.AsyncJsonWebTokenParser;
 import org.folio.sidecar.model.ScRoutingEntry;
 import org.folio.support.types.UnitTest;
 import org.junit.jupiter.api.AfterEach;
@@ -51,14 +52,14 @@ class KeycloakJwtFilterTest extends AbstractFilterTest {
 
   private static final String TEST_USER_ID = UUID.randomUUID().toString();
 
-  @Mock private JsonWebTokenParser jsonWebTokenParser;
+  @Mock private AsyncJsonWebTokenParser asyncJsonWebTokenParser;
   @Mock private HttpServerRequest request;
 
   @InjectMocks private KeycloakJwtFilter keycloakJwtFilter;
 
   @AfterEach
   void tearDown() {
-    verifyNoMoreInteractions(jsonWebTokenParser);
+    verifyNoMoreInteractions(asyncJsonWebTokenParser);
   }
 
   @Test
@@ -70,7 +71,7 @@ class KeycloakJwtFilterTest extends AbstractFilterTest {
     });
 
     var jsonWebToken = mock(JsonWebToken.class);
-    when(jsonWebTokenParser.parse(AUTH_TOKEN)).thenReturn(jsonWebToken);
+    when(asyncJsonWebTokenParser.parseAsync(AUTH_TOKEN)).thenReturn(Future.succeededFuture(jsonWebToken));
     when(jsonWebToken.getClaim(USER_ID_CLAIM)).thenReturn(TEST_USER_ID);
     when(jsonWebToken.getIssuer()).thenReturn("http://localhost:8080/auth/realms/" + TENANT_NAME);
 
@@ -96,7 +97,7 @@ class KeycloakJwtFilterTest extends AbstractFilterTest {
     });
 
     var jsonWebToken = mock(JsonWebToken.class);
-    when(jsonWebTokenParser.parse(AUTH_TOKEN)).thenReturn(jsonWebToken);
+    when(asyncJsonWebTokenParser.parseAsync(AUTH_TOKEN)).thenReturn(Future.succeededFuture(jsonWebToken));
     when(jsonWebToken.getIssuer()).thenReturn("http://localhost:8080/auth/realms/" + TENANT_NAME);
 
     var result = keycloakJwtFilter.applyFilter(routingContext);
@@ -121,7 +122,7 @@ class KeycloakJwtFilterTest extends AbstractFilterTest {
     });
 
     var jsonWebToken = mock(JsonWebToken.class);
-    when(jsonWebTokenParser.parse(AUTH_TOKEN)).thenReturn(jsonWebToken);
+    when(asyncJsonWebTokenParser.parseAsync(AUTH_TOKEN)).thenReturn(Future.succeededFuture(jsonWebToken));
     when(jsonWebToken.getClaim(USER_ID_CLAIM)).thenReturn(TEST_USER_ID);
     when(jsonWebToken.getIssuer()).thenReturn("http://localhost:8080/auth/realms/" + TENANT_NAME);
 
@@ -149,7 +150,7 @@ class KeycloakJwtFilterTest extends AbstractFilterTest {
     });
 
     var jsonWebToken = mock(JsonWebToken.class);
-    when(jsonWebTokenParser.parse(AUTH_TOKEN)).thenReturn(jsonWebToken);
+    when(asyncJsonWebTokenParser.parseAsync(AUTH_TOKEN)).thenReturn(Future.succeededFuture(jsonWebToken));
     when(jsonWebToken.getClaim(USER_ID_CLAIM)).thenReturn(TEST_USER_ID);
     when(jsonWebToken.getIssuer()).thenReturn("http://localhost:8080/auth/realms/" + TENANT_NAME);
 
@@ -178,7 +179,8 @@ class KeycloakJwtFilterTest extends AbstractFilterTest {
       when(request.headers()).thenReturn(requestHeaders);
     });
 
-    when(jsonWebTokenParser.parse(dummyToken)).thenThrow(new ParseException(INVALID_SEGMENTS_JWT_ERROR_MSG));
+    when(asyncJsonWebTokenParser.parseAsync(dummyToken)).thenReturn(Future.failedFuture(
+      new UnauthorizedException("Failed to parse JWT", new ParseException(INVALID_SEGMENTS_JWT_ERROR_MSG))));
 
     var result = keycloakJwtFilter.applyFilter(routingContext);
 
@@ -210,7 +212,7 @@ class KeycloakJwtFilterTest extends AbstractFilterTest {
     verify(routingContext, never()).put(anyString(), any());
 
     assertThat(requestHeaders.get(SYSTEM_TOKEN)).isEqualTo(systemToken);
-    verifyNoInteractions(jsonWebTokenParser);
+    verifyNoInteractions(asyncJsonWebTokenParser);
   }
 
   @Test
@@ -225,7 +227,8 @@ class KeycloakJwtFilterTest extends AbstractFilterTest {
       when(rc.get(SELF_REQUEST_KEY)).thenReturn(false);
     });
 
-    when(jsonWebTokenParser.parse(AUTH_TOKEN)).thenThrow(new ParseException("Invalid JWT"));
+    when(asyncJsonWebTokenParser.parseAsync(AUTH_TOKEN)).thenReturn(Future.failedFuture(
+      new UnauthorizedException("Failed to parse JWT", new ParseException("Invalid JWT"))));
 
     var result = keycloakJwtFilter.applyFilter(routingContext);
 
@@ -266,7 +269,7 @@ class KeycloakJwtFilterTest extends AbstractFilterTest {
       .isInstanceOf(BadRequestException.class)
       .hasMessage("X-Okapi-Token is not equal to Authorization token");
 
-    verifyNoInteractions(jsonWebTokenParser);
+    verifyNoInteractions(asyncJsonWebTokenParser);
   }
 
   @Test
@@ -277,7 +280,8 @@ class KeycloakJwtFilterTest extends AbstractFilterTest {
       when(request.headers()).thenReturn(requestHeaders);
     });
 
-    when(jsonWebTokenParser.parse(AUTH_TOKEN)).thenThrow(new ParseException("Failed to parse JWT, invalid offset"));
+    when(asyncJsonWebTokenParser.parseAsync(AUTH_TOKEN)).thenReturn(Future.failedFuture(
+      new UnauthorizedException("Failed to parse JWT", new ParseException("Failed to parse JWT, invalid offset"))));
 
     var result = keycloakJwtFilter.applyFilter(routingContext);
 
@@ -316,7 +320,8 @@ class KeycloakJwtFilterTest extends AbstractFilterTest {
       when(rc.get(SELF_REQUEST_KEY)).thenReturn(true);
     });
 
-    when(jsonWebTokenParser.parse(AUTH_TOKEN)).thenThrow(new ParseException("Invalid JWT"));
+    when(asyncJsonWebTokenParser.parseAsync(AUTH_TOKEN)).thenReturn(Future.failedFuture(
+      new UnauthorizedException("Failed to parse JWT", new ParseException("Invalid JWT"))));
 
     var result = keycloakJwtFilter.applyFilter(routingContext);